# improve modulo functions and number streams