--- a/fusequery/query/src/datasources/system/numbers_stream.rs
+++ b/fusequery/query/src/datasources/system/numbers_stream.rs
@@ -2,20 +2,14 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-use std::mem::ManuallyDrop;
-use std::mem::{self};
-use std::ptr::NonNull;
 use std::task::Context;
 use std::task::Poll;
 use std::usize;
 
-use common_arrow::arrow::array::ArrayData;
-use common_arrow::arrow::array::ArrayRef;
-use common_arrow::arrow::array::UInt64Array;
-use common_arrow::arrow::buffer::Buffer;
-use common_arrow::arrow::datatypes::DataType as ArrowDataType;
 use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
+use common_datavalues::AlignedVec;
+use common_datavalues::DFUInt64Array;
 use common_exception::Result;
 use common_streams::ProgressStream;
 use futures::stream::Stream;
@@ -93,36 +87,15 @@ impl NumbersStream {
         Ok(if current.begin == current.end {
             None
         } else {
-            let v = (current.begin..current.end).collect::<Vec<u64>>();
-            let mut me = ManuallyDrop::new(v);
-            let byte_size = mem::size_of::<u64>();
+            let mut av =
+                AlignedVec::with_capacity_len_aligned((current.end - current.begin) as usize);
 
-            unsafe {
-                let buffer = Buffer::from_raw_parts(
-                    NonNull::new(me.as_mut_ptr() as *mut u8).unwrap(),
-                    me.len() * byte_size,
-                    me.capacity() * byte_size,
-                );
-
-                let arr_data = ArrayData::builder(ArrowDataType::UInt64)
-                    .len(me.len())
-                    .offset(0)
-                    .add_buffer(buffer)
-                    .build();
-
-                let array = Arc::new(UInt64Array::from(arr_data)) as ArrayRef;
-                let block =
-                    DataBlock::create_by_array(self.schema.clone(), vec![array.into_series()]);
-                Some(block)
-            }
-
-            // let mut av = AlignedVec::with_capacity_aligned((current.end - current.begin) as usize);
-            // for val in current.begin..current.end {
-            //     av.push(val);
-            // }
-            // let series = DFUInt64Array::new_from_aligned_vec(av).into_series();
-            // let block = DataBlock::create_by_array(self.schema.clone(), vec![series]);
-            // Some(block)
+            av.iter_mut().enumerate().for_each(|(idx, num)| {
+                *num = current.begin + idx as u64;
+            });
+            let series = DFUInt64Array::new_from_aligned_vec(av).into_series();
+            let block = DataBlock::create_by_array(self.schema.clone(), vec![series]);
+            Some(block)
         })
     }
 }