I hereby agree to the terms of the CLA available at: https://datafuse.rs/policies/cla/

## Summary

- [x] use AlignVector to generate arrow buffers
- [x] improve UIntXX % Constant UInt8 cases
- [x] simplify some codes. 

The performance gains about 50%.

without snallocate
```
SELECT max(number),sum(number) FROM numbers_mt(1000000000) GROUP BY number % 3, number % 4, number % 5

60 rows in set. Elapsed: 5.058 sec. Processed 1.00 billion rows, 8.01 GB (197.71 million rows/s., 1.58 GB/s.)
```


with snallocate
```
SELECT max(number),sum(number) FROM numbers_mt(1000000000) GROUP BY number % 3, number % 4, number % 5

60 rows in set. Elapsed: 3.860 sec. Processed 1.00 billion rows, 8.01 GB (259.06 million rows/s., 2.07 GB/s.)
```


## Changelog

- Improvement


## Related Issues

Fixes #issue

## Test Plan

Unit Tests

Stateless Tests